+++ lib/Models/Catalog/Ows/CswCatalogGroup.ts
@@ -395,11 +395,10 @@ class CswStratum extends LoadableStratum(CswCatalogGroupTraits) {
       model = existingModel;
     }
     // Replace the stratum inherited from the parent group.
-    const stratum = CommonStrata.underride;
-    model.strata.delete(stratum);
-    model.setTrait(CommonStrata.underride, "name", metadataGroup.groupName);
+    model.strata.delete(CommonStrata.definition);
+    model.setTrait(CommonStrata.definition, "name", metadataGroup.groupName);
     model.setTrait(
-      CommonStrata.underride,
+      CommonStrata.definition,
       "members",
       filterOutUndefined([
         ...metadataGroup.children.map(
@@ -479,21 +478,25 @@ class CswStratum extends LoadableStratum(CswCatalogGroupTraits) {
         model = existingModel;
       }
       // Replace the stratum inherited from the parent group.
-      const stratum = CommonStrata.underride;
-      model.strata.delete(stratum);
-      model.setTrait(stratum, "name", record.title ?? record.identifier);
+      model.strata.delete(CommonStrata.definition);
+
+      model.setTrait(
+        CommonStrata.definition,
+        "name",
+        record.title ?? record.identifier
+      );
       const uri = acceptableUris[urlIndex];
-      model.setTrait(stratum, "url", uri.toString());
+      model.setTrait(CommonStrata.definition, "url", uri.toString());
 
       if (record.abstract) {
         model.setTrait(
-          stratum,
+          CommonStrata.definition,
           "description",
           toArray(record.abstract)?.join("\n\n")
         );
       } else if (record.description) {
         model.setTrait(
-          stratum,
+          CommonStrata.definition,
           "description",
           toArray(record.description)?.join("\n\n")
         );
@@ -515,9 +518,9 @@ class CswStratum extends LoadableStratum(CswCatalogGroupTraits) {
           .join("\n\n")
       });
 
-      model.setTrait(stratum, "info", infoSections);
+      model.setTrait(CommonStrata.definition, "info", infoSections);
 
-      model.setTrait(stratum, "metadataUrls", [
+      model.setTrait(CommonStrata.definition, "metadataUrls", [
         {
           title: i18next.t("models.csw.metadataURL"),
           url: new URI(
@@ -536,7 +539,9 @@ class CswStratum extends LoadableStratum(CswCatalogGroupTraits) {
       ]);
 
       if (legendUri) {
-        model.setTrait(stratum, "legends", [{ url: legendUri.toString() }]);
+        model.setTrait(CommonStrata.definition, "legends", [
+          { url: legendUri.toString() }
+        ]);
       }
 
       // If this is a WMS item, we MUST set `layers` trait to `uri.name`
@@ -544,7 +549,7 @@ class CswStratum extends LoadableStratum(CswCatalogGroupTraits) {
         if (!uri.name) {
           return;
         }
-        model.setTrait(stratum, "layers", uri.name);
+        model.setTrait(CommonStrata.definition, "layers", uri.name);
       }
 
       // Same with ArcGis MapServer
@@ -552,14 +557,7 @@ class CswStratum extends LoadableStratum(CswCatalogGroupTraits) {
         if (!uri.name) {
           return;
         }
-        model.setTrait(stratum, "layers", uri.name);
-      }
-
-      // Copy over itemProperties
-      if (this.catalogGroup.itemProperties !== undefined) {
-        Object.keys(this.catalogGroup.itemProperties).map((k: any) =>
-          model.setTrait(stratum, k, this.catalogGroup.itemProperties![k])
-        );
+        model.setTrait(CommonStrata.definition, "layers", uri.name);
       }
 
       return model;